+++ src/pages/Dashboard/useDashboardChainStatsMulticall.ts
@@ -2,12 +2,12 @@ import { zeroAddress } from "viem";
 
 import { getContract } from "config/contracts";
 import { getWhitelistedV1Tokens } from "config/tokens";
+import { useMulticall } from "lib/multicall/useMulticall";
+import { FREQUENT_MULTICALL_REFRESH_INTERVAL } from "lib/timeConstants";
 
 import GlpManager from "abis/GlpManager.json";
 import ReaderV2 from "abis/ReaderV2.json";
 import VaultV2 from "abis/VaultV2.json";
-import { useMulticall } from "lib/multicall/useMulticall";
-import { FREQUENT_MULTICALL_REFRESH_INTERVAL } from "lib/timeConstants";
 
 function buildDashboardRequest(chainId: number) {
   const gmxAddress = getContract(chainId, "GMX");
